chore: 移除所有 Citations 相关逻辑，为后续重构做准备
- Backend: 删除 lead_agent / general_purpose 中的 citations_format 与引用相关 reminder；artifacts 下载不再对 markdown 做 citation 清洗，统一走 FileResponse，保留 Response 用于二进制 inline
- Frontend: 删除 core/citations 模块、inline-citation、safe-citation-content；新增 MarkdownContent 仅做 Markdown 渲染；消息/artifact 预览与复制均使用原始 content
- i18n: 移除 citations 命名空间（loadingCitations、loadingCitationsWithCount）
- 技能与 demo: 措辞改为 references，demo 数据去掉 <citations> 块
- 文档: 更新 CLAUDE/AGENTS/README 描述，新增按文件 diff 的代码变更总结
