--- frontend/src/core/citations/index.ts
@@ -1,13 +1,0 @@
-export {
-  contentWithoutCitationsFromParsed,
-  extractDomainFromUrl,
-  isExternalUrl,
-  parseCitations,
-  removeAllCitations,
-  shouldShowCitationLoading,
-  syntheticCitationFromLink,
-} from "./utils";
-
-export { useParsedCitations } from "./use-parsed-citations";
-export type { UseParsedCitationsResult } from "./use-parsed-citations";
-export type { Citation, ParseCitationsResult } from "./utils";
